You can test run the test application for the single elimination.
The SingleElimination works perfectly but it might not work well if
there are ties. I would work on that before the next commit.

I found that the RoundRobin class contains some flaws. Would also
correct that in next commit

Tou can run the test applications for the roundRobin and
SingleElimination and possibly begin working on them

Everything is taking shape gradually